"#7 feat: add some type_trait utils class template"
- add remove_const template
- add remove_pointer template
- move is_pair from utility to type_traits file
